moved types to gql
