+++ packages/shared/index.ts
@@ -1,4 +1,3 @@
-export * from './datastore.js';
 export * from './integrations.js';
 export * from './templates.js';
 export * from './types.js';
