+++ packages/shared/types.ts
@@ -1,18 +1,3 @@
-import type { DataStore } from "./datastore.js";
-import { BaseConfig } from '@superglue/client';
-
-export type Context = {
-  datastore: DataStore;
-  orgId: string;
-};
-// Workflow related types
-export type ExecutionMode = "DIRECT" | "LOOP";
-
-export interface Metadata {
-  runId?: string;
-  orgId?: string;
-}
-
 export interface LogEntry {
   id: string;
   message: string;
